This is the PANEL2_BRANCH. It is for fixing all the misc brokennesses that
This is the PANEL2_BRANCH. It is for fixing all the misc brokennesses that
are causing people to have the panel hanging/segfaulting/etc. without
having a non-compilable trunk.

This does not currently compile. If you want to fix it so it does (i.e.
help convert it from using the old IDL:GNOME::Panel/1.0 interface to the
new one defined in panel2.idl) please tell jirka@5z.com and
sopwith@redhat.com.

This doesn't currently compile, but it will RSN :)